refactor: z-index 토큰화 및 중앙화(#77)
