+++ src/shared/components/baseModal/baseModal.css.ts
@@ -1,7 +1,6 @@
 import { style } from '@vanilla-extract/css';
-import { color } from '@/shared/styles/color.css';
 import { fadeIn } from '@/shared/styles/dialog.css';
-import { media } from '@/shared/styles';
+import { media, color, zIndex } from '@/shared/styles';
 export { overlay, hiddenTitle } from '@/shared/styles/dialog.css';
 
 export const content = style({
@@ -14,7 +13,7 @@ export const content = style({
   boxShadow: ' 0 0 10px 0 rgba(28, 35, 43, 0.20)',
   maxHeight: '90vh',
   overflowY: 'auto',
-  zIndex: 1001,
+  zIndex: zIndex.dialogContent,
   outline: 'none',
   animation: `${fadeIn} 0.3s ease-out`,
   width: '39.25rem',
