+++ src/shared/styles/index.ts
@@ -7,3 +7,4 @@ export * from './media';
 
 export { color } from './color.css';
 export { typography } from './typography.css';
+export { zIndex } from './zIndex.css';
